refactor(structure): move blender addon source files into scripts/ directory
