+++ assets/ComfyUI-PBR+mask.png
